organized the controllers folder and added hyper link to the login and signup button
