+++ controllers/index.js
@@ -1,12 +1,13 @@
 const router = require('express').Router();
 
 const homeRoutes = require('./homeRoutes');
-const projectRoute = require('./api/projectRoute');
-const apiRoutes = requrie('./api')
+const bucketlistRoutes = require('./bucketlistRoutes');
+const apiRoutes = require('./api')
 const dashboardRoutes = require('./dashboardRoutes')
 
 router.use('/', homeRoutes);
 router.use('/api', apiRoutes);
 router.use('/dashboard', dashboardRoutes);
+router.use("/bucketlist", bucketlistRoutes)
 
 module.exports = router;
