+++ views/homepage.handlebars
@@ -6,8 +6,8 @@
         <div class="col-lg-6 mx-auto">
           <p class="lead mb-4">Quickly find bucket list ideas. Create your own bucket list. Follow others as they journey to the end.</p>
           <div class="d-grid gap-2 d-sm-flex justify-content-center">
-            <a class="btn btn-primary btn-lg px-4">Login</a>
-            <a class="btn btn-primary btn-lg px-4">Sign up</a>
+            <a class="btn btn-primary btn-lg px-4" href="/login">Login</a>
+            <a class="btn btn-primary btn-lg px-4" href="/signup">Sign up</a>
           </div>
           <div class="d-grid gap-2 d-sm-flex justify-content-sm-center">
             <button id="homeIdea" type="button" class="btn btn-success btn-lg px-4 my-4">Get an idea</button>
